+++ src/rules/UnnecessaryGString.js
@@ -1,8 +1,10 @@
 // Unnecessary Groovy String
 
-const { getVariableRange } = require("../utils");
+const { getVariable, getVariableRange, findRangeBetweenStrings } = require("../utils");
 
 const rule = {
+    scope: "file",
+    unitary: true,
     variables: [
         {
             name: "STRING",
@@ -11,15 +13,34 @@ const rule = {
     ],
     range: {
         type: "function",
-        func: (errLine, errItem, evaluatedVars) => {
-            return getVariableRange(errLine, evaluatedVars, "STRING", errItem);
+        func: (errLine, errItem, evaluatedVars, allLines) => {
+            // Single line range
+            const singleLineRange = getVariableRange(errLine, evaluatedVars, "STRING", errItem);
+            if (singleLineRange.start.character > -1) {
+                return singleLineRange;
+            } else {
+                return findRangeBetweenStrings(allLines, errItem, '"""', '"""');
+            }
         }
     },
     fix: {
         label: "Replace double quotes by single quotes",
-        type: "replaceString",
-        before: '"{{STRING}}"',
-        after: "'{{STRING}}'"
+        type: "function",
+        func: (allLines, variables) => {
+            const lineNumber = getVariable(variables, "lineNb", { mandatory: true });
+            const range = getVariable(variables, "range", { mandatory: true });
+            const str = getVariable(variables, "STRING", { mandatory: true });
+            // Single line replacement: replace " by '
+            if (range.start.line === range.end.line) {
+                allLines[lineNumber] = allLines[lineNumber].replace(`"${str}"`, `'${str}'`);
+            }
+            // Multiline replacement: replace """ by '''
+            else {
+                allLines[range.start.line - 1] = allLines[range.start.line - 1].replace(`"""`, `'''`);
+                allLines[range.end.line - 1] = allLines[range.end.line - 1].replace(`"""`, `'''`);
+            }
+            return allLines;
+        }
     },
     tests: [
         {
@@ -52,6 +73,28 @@ String str = 'lelamanul' + "\\n\\r\\n" + "titi\\n" + "\\n\\r" + "lelamanul\\nwes
 `,
             sourceAfter: `
 String str = 'lelamanul' + '\\n\\r\\n' + 'titi\\n' + '\\n\\r' + 'lelamanul\\nwesh'
+`
+        },
+        {
+            sourceBefore: `
+def b = """
+I am a string
+"""
+`,
+            sourceAfter: `
+def b = '''
+I am a string
+'''
+`
+        },
+        {
+            sourceBefore: `
+def b = """
+I am a string"""
+`,
+            sourceAfter: `
+def b = '''
+I am a string'''
 `
         }
     ]